Cancel booking  confirmation modal

+++ src/components/UserSlot/UserSlot.js
@@ -1,14 +1,16 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { useDispatch } from "react-redux";
+
+import { withConfirm } from "../withConfirm";
 
 import { deleteUsersBooking } from "../../redux/userBookings/userBookingsUtils";
 
-export const UserSlot = ({ data: { date, hours, _id }, number }) => {
+const UserSlot = ({ data: { date, hours, _id }, number, setOpen, confirm }) => {
   const dispatch = useDispatch();
 
-  const deleteSlot = () => {
-    dispatch(deleteUsersBooking(date, _id));
-  };
+  useEffect(() => {
+    if (confirm) dispatch(deleteUsersBooking(date, _id));
+  }, [_id, confirm, date, dispatch]);
 
   return (
     <div className="user-bookings__slot">
@@ -18,10 +20,13 @@ export const UserSlot = ({ data: { date, hours, _id }, number }) => {
       </div>
       <div className="user-bookings__slot hours">{hours}</div>
       <div className="user-bookings__slot cancel">
-        <button className="cancel" onClick={async () => deleteSlot()}>
+        <button className="cancel" onClick={async () => setOpen(true)}>
           X
         </button>
       </div>
     </div>
   );
 };
+
+const userSlotWithConfirm = withConfirm(UserSlot);
+export { userSlotWithConfirm as UserSlot };
